// frontend/src/app/common/register-dto.ts
export class RegisterDTO {
        id?:Number;
        username?: string;
        password?: string;
        confirmPassword?: string;
        firstName?: string;
        lastName?: string;
        email?: string;
        cityName?: string;
        streetName?: string;
        cm?: Number;
        kg?: Number;
        bornDate?: Date;
        gender?: boolean;
        role?: boolean;
}